Introduce a new `io_buffer` module.
It abstracts buffers involved in IO. This is in preparation for allowing
drivers to implement `write_iter` and `read_iter`.

No behaviour change is intended, this is a pure refactor.
