+++ drivers/android/transaction.rs
@@ -2,7 +2,10 @@
 
 use alloc::sync::Arc;
 use core::sync::atomic::{AtomicBool, Ordering};
-use kernel::{bindings, linked_list::Links, prelude::*, sync::Ref, user_ptr::UserSlicePtrWriter};
+use kernel::{
+    bindings, io_buffer::IoBufferWriter, linked_list::Links, prelude::*, sync::Ref,
+    user_ptr::UserSlicePtrWriter,
+};
 
 use crate::{
     defs::*,
